feat(table): subcomponents table and auto table BREAKING CHANGE: AutoTable replaces DataTable, Table replaces ZebraTable
