+++ packages/table/src/table-row-column.tsx
@@ -1,20 +1,14 @@
-import { DefaultProps, cx } from '@sk-web-gui/utils';
+import { cx } from '@sk-web-gui/utils';
 import React from 'react';
 
-interface ITableRowColumnProps extends DefaultProps {
-  children: JSX.Element | JSX.Element[] | string;
-}
+export const TableRowColumn = React.forwardRef<HTMLTableCellElement, React.ComponentPropsWithoutRef<'td'>>(
+  (props, ref) => {
+    const { children, className, ...rest } = props;
 
-export interface TableRowColumnProps
-  extends Omit<React.HTMLAttributes<HTMLTableCellElement>, 'color' | 'children'>,
-    ITableRowColumnProps {}
-
-export const TableRowColumn = React.forwardRef<HTMLTableCellElement, TableRowColumnProps>((props, ref) => {
-  const { children, className, ...rest } = props;
-
-  return (
-    <td ref={ref} className={cx('sk-table-tbody-td sk-table-tbody-manualtd-content', className)} {...rest}>
-      {children}
-    </td>
-  );
-});
+    return (
+      <td ref={ref} className={cx('sk-table-tbody-td sk-table-tbody-manualtd-content', className)} {...rest}>
+        {children}
+      </td>
+    );
+  }
+);
