+++ packages/table/src/table-sort-button.tsx
@@ -1,23 +1,17 @@
 import { Icon } from '@sk-web-gui/icon';
-import { DefaultProps, cx } from '@sk-web-gui/utils';
+import { DefaultProps } from '@sk-web-gui/utils';
 import React from 'react';
-import { sortMode } from './table';
+import { sortMode } from './auto-table';
 
-interface ITableSortButtonProps extends DefaultProps {
+export interface ITableSortButtonProps extends DefaultProps, React.ComponentPropsWithoutRef<'button'> {
   children: JSX.Element | JSX.Element[] | string;
   isActive: boolean;
   sortOrder: sortMode.ASC | sortMode.DESC | null;
-  isColumnSortable: boolean;
-  screenReaderOnly: boolean;
 }
 
-export interface TableSortButtonProps
-  extends Omit<React.HTMLAttributes<HTMLButtonElement>, 'color' | 'children'>,
-    ITableSortButtonProps {}
-
-export const TableSortButton = React.forwardRef<HTMLButtonElement, TableSortButtonProps>((props, ref) => {
-  const { sortOrder, children, isActive, isColumnSortable, screenReaderOnly, ...rest } = props;
-  return isColumnSortable ? (
+export const TableSortButton = React.forwardRef<HTMLButtonElement, ITableSortButtonProps>((props, ref) => {
+  const { sortOrder, children, isActive, ...rest } = props;
+  return (
     <button ref={ref} {...rest} className="sk-table-sortbutton">
       {children}
       <div className="sk-table-sortbutton-icon">
@@ -34,9 +28,5 @@ export const TableSortButton = React.forwardRef<HTMLButtonElement, TableSortButt
         )}
       </div>
     </button>
-  ) : (
-    <span ref={ref} {...rest} className={cx('sk-table-sortbutton', `${screenReaderOnly ? `sr-only` : ``}`)}>
-      {children}
-    </span>
   );
 });
